aether-backend/requirements.txt
# Aether Backend - Production Dependencies
# Install with: pip install -r requirements.txt

# Core Framework
fastapi==0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# HTTP Client (compatible with MCP)
httpx>=0.26.0

# Database
asyncpg==0.29.0
aiosqlite==0.19.0
sqlalchemy[asyncio]==2.0.25
psycopg[binary]==3.1.18
psycopg-pool==3.2.4

# Settings Management (compatible with MCP)
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv==1.0.0

# Configuration
toml==0.10.2
pyyaml==6.0.1

# Security
cryptography==42.0.0
bcrypt==4.1.2

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Rate Limiting
slowapi==0.1.9

# WebSockets
websockets==12.0
python-socketio==5.11.0

# Data Validation
email-validator==2.1.0

# Utilities
python-json-logger==2.0.7
tenacity==8.2.3

# MCP - Model Context Protocol
mcp==1.20.0

# Open Interpreter Dependencies (from pyproject.toml)
shortuuid>=1.0.0
litellm>=1.0.0
rich>=13.0.0
inquirer>=3.1.3
astor>=0.8.1
git-python>=1.0.3
six>=1.16.0
tokentrim>=0.1.13
wget>=3.2
html2image>=2.0.4
send2trash>=1.8.2
ipykernel>=6.26.0
jupyter-client>=8.6.0
matplotlib>=3.8.2
tiktoken>=0.7.0
platformdirs>=4.2.0
google-generativeai>=0.7.1
pyperclip>=1.9.0
yaspin>=3.0.2
html2text>=2024.2.26
selenium>=4.24.0
webdriver-manager>=4.0.2
anthropic>=0.37.1
pyautogui>=0.9.54
typer>=0.12.5
scikit-learn>=1.0.0
opencv-python>=4.8.0
janus>=1.0.0

# Open Interpreter (from local path - managed separately)
# interpreter - installed from backend/open-interpreter

# TTS Integration
# RealtimeTTS - installed from services/realtime-tts
# Core dependencies for basic TTS engines
stream2sentence==0.3.1
resampy==0.4.3
pydub==0.25.1
pyttsx3==2.90
gtts==2.4.0
edge-tts==6.1.9

# OCR Integration
# Chandra - installed from services/chandra
# Core dependencies for OCR
filetype==1.2.0
markdownify==1.1.0
pypdfium2>=4.30.0
qwen-vl-utils>=0.0.14
accelerate>=1.11.0
transformers>=4.57.1  # Required for Qwen3VL support

# Optional: Performance
orjson==3.9.12
ujson==5.9.0

# Optional: Production server
gunicorn==21.2.0

# Development tools (commented out for production)
# black==24.1.1
# ruff==0.1.14
# mypy==1.8.0

